Testing on Mesa, seems to work fine
Mesa does seem to add some extra padding after the last element in the
block but I don't think this is an issue since nothing comes after it.
Also seems to not generate names/indices for unused uniforms in the
block but does allocate space for them, so now we pretend to use the
members for testing.