Testing a more complex model revealed some bugs in model loading
Was using the wrong indexing into the model which would work for models
without any shared vertices but fails for ones that do share vertices
since we go way out of bounds. Now use the proper loop and vbo size.

Also found a bug in the sizer for layout types where we would return the
size of the previous + sizeof the type when really we should just return
the size of the types passed and only use the previous size to compute
offsets to pad ourselves properly. This was causing the tex coords (the
last glm::vec3) to be at offset 36 when they should have been at 24