Starting to make progress on more generic render batch support
Added support for writing entire blocks of elements to the interleaved
buffer by simply passing in a tuple and an index to write it too.

Also adding a template GLEnum type lookup to associate float with
GL_FLOAT for example. This will be used to implement the testing
set_attrib_index stuff in Test in main fully in RenderBatch to handle
setting up the attribute indices